Working on nice features for that game logic
